🏗️ Refactor flows module
* Split the flows module into many sub-modules
* Rename {Distribution,Transform}Module to {Distribution,Transform}Factory
* Spread the bisection bounds of monotonic transformations
* New rotation transformation (RotationTransform)
* New gaussianization transformation (GaussianizationTransform)
* New gaussianization flow (GF)
