+++ tests/test_flows.py
@@ -8,37 +8,43 @@ from torch import randn
 from zuko.flows import *
 
 
+torch.set_default_dtype(torch.float64)
+
+
 def test_flows(tmp_path):
-    flows = [
-        GMM(3, 5),
-        MAF(3, 5),
-        NSF(3, 5),
-        SOSPF(3, 5),
-        NAF(3, 5),
-        UNAF(3, 5),
-        GCF(3, 5),
-        CNF(3, 5),
+    Fs = [
+        GMM,
+        NICE,
+        MAF,
+        NSF,
+        SOSPF,
+        NAF,
+        UNAF,
+        CNF,
+        GF,
     ]
 
-    for flow in flows:
+    for F in Fs:
+        flow = F(3, 5)
+
         # Evaluation of log_prob
         x, c = randn(256, 3), randn(5)
         log_p = flow(c).log_prob(x)
 
-        assert log_p.shape == (256,), flow
-        assert log_p.requires_grad, flow
+        assert log_p.shape == (256,), F
+        assert log_p.requires_grad, F
 
         flow.zero_grad(set_to_none=True)
         loss = -log_p.mean()
         loss.backward()
 
         for p in flow.parameters():
-            assert p.grad is not None, flow
+            assert p.grad is not None, F
 
         # Sampling
         x = flow(c).sample((32,))
 
-        assert x.shape == (32, 3), flow
+        assert x.shape == (32, 3), F
 
         # Reparameterization trick
         if flow(c).has_rsample:
@@ -49,15 +55,15 @@ def test_flows(tmp_path):
             loss.backward()
 
             for p in flow.parameters():
-                assert p.grad is not None, flow
+                assert p.grad is not None, F
 
         # Invertibility
-        if isinstance(flow, FlowModule):
+        if isinstance(flow, Flow):
             x, c = randn(256, 3), randn(256, 5)
             t = flow(c).transform
             z = t.inv(t(x))
 
-            assert torch.allclose(x, z, atol=1e-4), flow
+            assert torch.allclose(x, z, atol=1e-4), F
 
         # Saving
         torch.save(flow, tmp_path / 'flow.pth')
@@ -72,21 +78,21 @@ def test_flows(tmp_path):
         torch.manual_seed(seed)
         log_p_bis = flow_bis(c).log_prob(x)
 
-        assert torch.allclose(log_p, log_p_bis), flow
+        assert torch.allclose(log_p, log_p_bis), F
 
         # Printing
-        assert repr(flow), flow
+        assert repr(flow), F
 
 
 def test_triangular_transforms():
     Ts = [
         ElementWiseTransform,
+        GeneralCouplingTransform,
         MaskedAutoregressiveTransform,
         partial(MaskedAutoregressiveTransform, passes=2),
         NeuralAutoregressiveTransform,
         partial(NeuralAutoregressiveTransform, passes=2),
         UnconstrainedNeuralAutoregressiveTransform,
-        GeneralCouplingTransform,
     ]
 
     for T in Ts:
@@ -97,16 +103,16 @@ def test_triangular_transforms():
 
         assert y.shape == x.shape, t
         assert y.requires_grad, t
-        assert torch.allclose(t().inv(y), x, atol=1e-4), t
+        assert torch.allclose(t().inv(y), x, atol=1e-4), T
 
         # With context
         t = T(3, 5)
         x, c = randn(256, 3), randn(5)
         y = t(c)(x)
 
-        assert y.shape == x.shape, t
-        assert y.requires_grad, t
-        assert torch.allclose(t(c).inv(y), x, atol=1e-4), t
+        assert y.shape == x.shape, T
+        assert y.requires_grad, T
+        assert torch.allclose(t(c).inv(y), x, atol=1e-4), T
 
         # Jacobian
         t = T(7)
@@ -116,5 +122,5 @@ def test_triangular_transforms():
         J = torch.autograd.functional.jacobian(t(), x)
         ladj = torch.linalg.slogdet(J).logabsdet
 
-        assert torch.allclose(t().log_abs_det_jacobian(x, y), ladj, atol=1e-4), t
-        assert torch.allclose(J.diag().abs().log().sum(), ladj, atol=1e-4), t
+        assert torch.allclose(t().log_abs_det_jacobian(x, y), ladj, atol=1e-4), T
+        assert torch.allclose(J.diag().abs().log().sum(), ladj, atol=1e-4), T
