+++ zuko/flows/autoregressive.py
@@ -1,0 +1,225 @@
+r"""Autoregressive flows and transformations."""
+
+__all__ = [
+    'MaskedAutoregressiveTransform',
+    'MAF',
+]
+
+import torch
+import torch.nn as nn
+
+from functools import partial
+from math import ceil, prod
+from torch import Tensor, LongTensor, Size
+from torch.distributions import Transform
+from typing import *
+
+from .core import *
+from ..distributions import *
+from ..transforms import *
+from ..nn import MaskedMLP
+from ..utils import broadcast, unpack
+
+
+class MaskedAutoregressiveTransform(TransformFactory):
+    r"""Creates a masked autoregressive transformation.
+
+    References:
+        | Masked Autoregressive Flow for Density Estimation (Papamakarios et al., 2017)
+        | https://arxiv.org/abs/1705.07057
+
+    Arguments:
+        features: The number of features.
+        context: The number of context features.
+        passes: The number of sequential passes for the inverse transformation. If
+            :py:`None`, use the number of features instead, making the transformation
+            fully autoregressive. Coupling corresponds to :py:`passes=2`.
+        order: The feature ordering. If :py:`None`, use :py:`range(features)` instead.
+        univariate: The univariate transformation constructor.
+        shapes: The shapes of the univariate transformation parameters.
+        kwargs: Keyword arguments passed to :class:`zuko.nn.MaskedMLP`.
+
+    Example:
+        >>> t = MaskedAutoregressiveTransform(3, 4)
+        >>> t
+        MaskedAutoregressiveTransform(
+          (base): MonotonicAffineTransform()
+          (order): [0, 1, 2]
+          (hyper): MaskedMLP(
+            (0): MaskedLinear(in_features=7, out_features=64, bias=True)
+            (1): ReLU()
+            (2): MaskedLinear(in_features=64, out_features=64, bias=True)
+            (3): ReLU()
+            (4): MaskedLinear(in_features=64, out_features=6, bias=True)
+          )
+        )
+        >>> x = torch.randn(3)
+        >>> x
+        tensor([-0.9485,  1.5290,  0.2018])
+        >>> c = torch.randn(4)
+        >>> y = t(c)(x)
+        >>> t(c).inv(y)
+        tensor([-0.9485,  1.5290,  0.2018])
+    """
+
+    def __init__(
+        self,
+        features: int,
+        context: int = 0,
+        passes: int = None,
+        order: LongTensor = None,
+        univariate: Callable[..., Transform] = MonotonicAffineTransform,
+        shapes: Sequence[Size] = ((), ()),
+        **kwargs,
+    ):
+        super().__init__()
+
+        # Univariate transformation
+        self.univariate = univariate
+        self.shapes = shapes
+        self.total = sum(prod(s) for s in shapes)
+
+        # Adjacency
+        self.register_buffer('order', None)
+
+        if passes is None:
+            passes = features
+
+        if order is None:
+            order = torch.arange(features)
+        else:
+            order = torch.as_tensor(order)
+
+        self.passes = min(max(passes, 1), features)
+        self.order = torch.div(order, ceil(features / self.passes), rounding_mode='floor')
+
+        in_order = torch.cat((self.order, torch.full((context,), -1)))
+        out_order = torch.repeat_interleave(self.order, self.total)
+        adjacency = out_order[:, None] > in_order
+
+        # Hyper network
+        self.hyper = MaskedMLP(adjacency, **kwargs)
+
+    def extra_repr(self) -> str:
+        base = self.univariate(*map(torch.randn, self.shapes))
+        order = self.order.tolist()
+
+        if len(order) > 10:
+            order = order[:5] + [...] + order[-5:]
+            order = str(order).replace('Ellipsis', '...')
+
+        return '\n'.join([
+            f'(base): {base}',
+            f'(order): {order}',
+        ])
+
+    def meta(self, c: Tensor, x: Tensor) -> Transform:
+        if c is not None:
+            x = torch.cat(broadcast(x, c, ignore=1), dim=-1)
+
+        phi = self.hyper(x)
+        phi = phi.unflatten(-1, (-1, self.total))
+        phi = unpack(phi, self.shapes)
+
+        return DependentTransform(self.univariate(*phi), 1)
+
+    def forward(self, c: Tensor = None) -> Transform:
+        return AutoregressiveTransform(partial(self.meta, c), self.passes)
+
+
+class MAF(Flow):
+    r"""Creates a masked autoregressive flow (MAF).
+
+    References:
+        | Masked Autoregressive Flow for Density Estimation (Papamakarios et al., 2017)
+        | https://arxiv.org/abs/1705.07057
+
+    Arguments:
+        features: The number of features.
+        context: The number of context features.
+        transforms: The number of autoregressive transformations.
+        randperm: Whether features are randomly permuted between transformations or not.
+            If :py:`False`, features are in ascending (descending) order for even
+            (odd) transformations.
+        kwargs: Keyword arguments passed to :class:`MaskedAutoregressiveTransform`.
+
+    Example:
+        >>> flow = MAF(3, 4, transforms=3)
+        >>> flow
+        MAF(
+          (transforms): ModuleList(
+            (0): MaskedAutoregressiveTransform(
+              (base): MonotonicAffineTransform()
+              (order): [0, 1, 2]
+              (hyper): MaskedMLP(
+                (0): MaskedLinear(in_features=7, out_features=64, bias=True)
+                (1): ReLU()
+                (2): MaskedLinear(in_features=64, out_features=64, bias=True)
+                (3): ReLU()
+                (4): MaskedLinear(in_features=64, out_features=6, bias=True)
+              )
+            )
+            (1): MaskedAutoregressiveTransform(
+              (base): MonotonicAffineTransform()
+              (order): [2, 1, 0]
+              (hyper): MaskedMLP(
+                (0): MaskedLinear(in_features=7, out_features=64, bias=True)
+                (1): ReLU()
+                (2): MaskedLinear(in_features=64, out_features=64, bias=True)
+                (3): ReLU()
+                (4): MaskedLinear(in_features=64, out_features=6, bias=True)
+              )
+            )
+            (2): MaskedAutoregressiveTransform(
+              (base): MonotonicAffineTransform()
+              (order): [0, 1, 2]
+              (hyper): MaskedMLP(
+                (0): MaskedLinear(in_features=7, out_features=64, bias=True)
+                (1): ReLU()
+                (2): MaskedLinear(in_features=64, out_features=64, bias=True)
+                (3): ReLU()
+                (4): MaskedLinear(in_features=64, out_features=6, bias=True)
+              )
+            )
+          )
+          (base): DiagNormal(loc: torch.Size([3]), scale: torch.Size([3]))
+        )
+        >>> c = torch.randn(4)
+        >>> x = flow(c).sample()
+        >>> x
+        tensor([-1.7154, -0.4401,  0.7505])
+        >>> flow(c).log_prob(x)
+        tensor(-4.4630, grad_fn=<AddBackward0>)
+    """
+
+    def __init__(
+        self,
+        features: int,
+        context: int = 0,
+        transforms: int = 3,
+        randperm: bool = False,
+        **kwargs,
+    ):
+        orders = [
+            torch.arange(features),
+            torch.flipud(torch.arange(features)),
+        ]
+
+        transforms = [
+            MaskedAutoregressiveTransform(
+                features=features,
+                context=context,
+                order=torch.randperm(features) if randperm else orders[i % 2],
+                **kwargs,
+            )
+            for i in range(transforms)
+        ]
+
+        base = Unconditional(
+            DiagNormal,
+            torch.zeros(features),
+            torch.ones(features),
+            buffer=True,
+        )
+
+        super().__init__(transforms, base)
