+++ zuko/flows/neural.py
@@ -1,0 +1,287 @@
+r"""Neural flows and transformations."""
+
+__all__ = [
+    'NeuralAutoregressiveTransform',
+    'NAF',
+    'UnconstrainedNeuralAutoregressiveTransform',
+    'UNAF',
+]
+
+import torch
+import torch.nn as nn
+
+from functools import partial
+from torch import Tensor
+from torch.distributions import Transform
+from typing import *
+
+from .autoregressive import *
+from .core import *
+from ..distributions import *
+from ..transforms import *
+from ..nn import MLP, MonotonicMLP
+from ..utils import broadcast
+
+
+class NeuralAutoregressiveTransform(MaskedAutoregressiveTransform):
+    r"""Creates a neural autoregressive transformation.
+
+    The monotonic neural network is parametrized by its internal positive weights,
+    which are independent of the features and context. To modulate its behavior, it
+    receives as input a signal that is autoregressively dependent on the features
+    and context.
+
+    References:
+        | Neural Autoregressive Flows (Huang et al., 2018)
+        | https://arxiv.org/abs/1804.00779
+
+    Arguments:
+        features: The number of features.
+        context: The number of context features.
+        signal: The number of signal features of the monotonic network.
+        network: Keyword arguments passed to :class:`zuko.nn.MonotonicMLP`.
+        kwargs: Keyword arguments passed to :class:`MaskedAutoregressiveTransform`.
+
+    Example:
+        >>> t = NeuralAutoregressiveTransform(3, 4)
+        >>> t
+        NeuralAutoregressiveTransform(
+          (base): MonotonicTransform()
+          (order): [0, 1, 2]
+          (hyper): MaskedMLP(
+            (0): MaskedLinear(in_features=7, out_features=64, bias=True)
+            (1): ReLU()
+            (2): MaskedLinear(in_features=64, out_features=64, bias=True)
+            (3): ReLU()
+            (4): MaskedLinear(in_features=64, out_features=24, bias=True)
+          )
+          (network): MonotonicMLP(
+            (0): MonotonicLinear(in_features=9, out_features=64, bias=True)
+            (1): TwoWayELU(alpha=1.0)
+            (2): MonotonicLinear(in_features=64, out_features=64, bias=True)
+            (3): TwoWayELU(alpha=1.0)
+            (4): MonotonicLinear(in_features=64, out_features=1, bias=True)
+          )
+        )
+        >>> x = torch.randn(3)
+        >>> x
+        tensor([-2.3267,  1.4581, -1.6776])
+        >>> c = torch.randn(4)
+        >>> y = t(c)(x)
+        >>> t(c).inv(y)
+        tensor([-2.3267,  1.4581, -1.6776])
+    """
+
+    def __init__(
+        self,
+        features: int,
+        context: int = 0,
+        signal: int = 8,
+        network: Dict[str, Any] = {},
+        **kwargs,
+    ):
+        super().__init__(
+            features=features,
+            context=context,
+            univariate=self.univariate,
+            shapes=[(signal,)],
+            **kwargs,
+        )
+
+        self.network = MonotonicMLP(1 + signal, 1, **network)
+
+    def f(self, signal: Tensor, x: Tensor) -> Tensor:
+        return self.network(
+            torch.cat(broadcast(x[..., None], signal, ignore=1), dim=-1)
+        ).squeeze(dim=-1)
+
+    def univariate(self, signal: Tensor) -> Transform:
+        return MonotonicTransform(
+            f=partial(self.f, signal),
+            phi=(signal, *self.network.parameters()),
+        )
+
+
+class NAF(Flow):
+    r"""Creates a neural autoregressive flow (NAF).
+
+    References:
+        | Neural Autoregressive Flows (Huang et al., 2018)
+        | https://arxiv.org/abs/1804.00779
+
+    Arguments:
+        features: The number of features.
+        context: The number of context features.
+        transforms: The number of autoregressive transformations.
+        randperm: Whether features are randomly permuted between transformations or not.
+            If :py:`False`, features are in ascending (descending) order for even
+            (odd) transformations.
+        unconstrained: Whether to use unconstrained or regular monotonic networks.
+        kwargs: Keyword arguments passed to :class:`NeuralAutoregressiveTransform`.
+    """
+
+    def __init__(
+        self,
+        features: int,
+        context: int = 0,
+        transforms: int = 3,
+        randperm: bool = False,
+        **kwargs,
+    ):
+        orders = [
+            torch.arange(features),
+            torch.flipud(torch.arange(features)),
+        ]
+
+        transforms = [
+            NeuralAutoregressiveTransform(
+                features=features,
+                context=context,
+                order=torch.randperm(features) if randperm else orders[i % 2],
+                **kwargs,
+            )
+            for i in range(transforms)
+        ]
+
+        base = Unconditional(
+            DiagNormal,
+            torch.zeros(features),
+            torch.ones(features),
+            buffer=True,
+        )
+
+        super().__init__(transforms, base)
+
+
+class UnconstrainedNeuralAutoregressiveTransform(MaskedAutoregressiveTransform):
+    r"""Creates an unconstrained neural autoregressive transformation.
+
+    The integrand neural network is parametrized by its internal weights, which are
+    independent of the features and context. To modulate its behavior, it receives as
+    input a signal that is autoregressively dependent on the features and context. The
+    integration constant has the same dependencies as the signal.
+
+    References:
+        | Unconstrained Monotonic Neural Networks (Wehenkel et al., 2019)
+        | https://arxiv.org/abs/1908.05164
+
+    Arguments:
+        features: The number of features.
+        context: The number of context features.
+        signal: The number of signal features of the integrand network.
+        network: Keyword arguments passed to :class:`zuko.nn.MLP`.
+        kwargs: Keyword arguments passed to :class:`MaskedAutoregressiveTransform`.
+
+    Example:
+        >>> t = UnconstrainedNeuralAutoregressiveTransform(3, 4)
+        >>> t
+        UnconstrainedNeuralAutoregressiveTransform(
+          (base): UnconstrainedMonotonicTransform()
+          (order): [0, 1, 2]
+          (hyper): MaskedMLP(
+            (0): MaskedLinear(in_features=7, out_features=64, bias=True)
+            (1): ReLU()
+            (2): MaskedLinear(in_features=64, out_features=64, bias=True)
+            (3): ReLU()
+            (4): MaskedLinear(in_features=64, out_features=27, bias=True)
+          )
+          (integrand): MLP(
+            (0): Linear(in_features=9, out_features=64, bias=True)
+            (1): ELU(alpha=1.0)
+            (2): Linear(in_features=64, out_features=64, bias=True)
+            (3): ELU(alpha=1.0)
+            (4): Linear(in_features=64, out_features=1, bias=True)
+            (5): Softplus(beta=1, threshold=20)
+          )
+        )
+        >>> x = torch.randn(3)
+        >>> x
+        tensor([-0.0103, -1.0871, -0.0667])
+        >>> c = torch.randn(4)
+        >>> y = t(c)(x)
+        >>> t(c).inv(y)
+        tensor([-0.0103, -1.0871, -0.0667])
+    """
+
+    def __init__(
+        self,
+        features: int,
+        context: int = 0,
+        signal: int = 8,
+        network: Dict[str, Any] = {},
+        **kwargs,
+    ):
+        super().__init__(
+            features=features,
+            context=context,
+            univariate=self.univariate,
+            shapes=[(signal,), ()],
+            **kwargs,
+        )
+
+        network.setdefault('activation', nn.ELU)
+
+        self.integrand = MLP(1 + signal, 1, **network)
+        self.integrand.append(nn.Softplus())
+
+    def g(self, signal: Tensor, x: Tensor) -> Tensor:
+        return self.integrand(
+            torch.cat(broadcast(x[..., None], signal, ignore=1), dim=-1)
+        ).squeeze(dim=-1)
+
+    def univariate(self, signal: Tensor, constant: Tensor) -> Transform:
+        return UnconstrainedMonotonicTransform(
+            g=partial(self.g, signal),
+            C=constant,
+            phi=(signal, *self.integrand.parameters()),
+        )
+
+
+class UNAF(Flow):
+    r"""Creates an unconstrained neural autoregressive flow (UNAF).
+
+    References:
+        | Unconstrained Monotonic Neural Networks (Wehenkel et al., 2019)
+        | https://arxiv.org/abs/1908.05164
+
+    Arguments:
+        features: The number of features.
+        context: The number of context features.
+        transforms: The number of autoregressive transformations.
+        randperm: Whether features are randomly permuted between transformations or not.
+            If :py:`False`, features are in ascending (descending) order for even
+            (odd) transformations.
+        kwargs: Keyword arguments passed to :class:`UnconstrainedNeuralAutoregressiveTransform`.
+    """
+
+    def __init__(
+        self,
+        features: int,
+        context: int = 0,
+        transforms: int = 3,
+        randperm: bool = False,
+        **kwargs,
+    ):
+        orders = [
+            torch.arange(features),
+            torch.flipud(torch.arange(features)),
+        ]
+
+        transforms = [
+            UnconstrainedNeuralAutoregressiveTransform(
+                features=features,
+                context=context,
+                order=torch.randperm(features) if randperm else orders[i % 2],
+                **kwargs,
+            )
+            for i in range(transforms)
+        ]
+
+        base = Unconditional(
+            DiagNormal,
+            torch.zeros(features),
+            torch.ones(features),
+            buffer=True,
+        )
+
+        super().__init__(transforms, base)
